print packet number as unsigned int

+++ newbasic/oncsSub_idgl1v0.cc
@@ -256,7 +256,7 @@ void oncsSub_idgl1v0::dump(std::ostream &os)
       return;
     }
   
-  os << "packet nr:       " << iValue(0) << endl;
+  os << "packet nr:       " << (unsigned int) iValue(0) << endl;
   os << "Beam Clock:      " << "0x" << hex << lValue(0, "BCO") << dec  << "   "  << lValue(0, "BCO") << endl;
   os << "Trigger Input:   " << "0x" << hex <<  lValue(0, "TriggerInput")  << dec << "   " << lValue(0, "TriggerInput") << endl;
   os << "Trigger Vector:  " << "0x" << hex <<  lValue(0, "TriggerVector") << dec << "   " << lValue(0, "TriggerVector") << endl;
